New file basic_alogirthms.R which will contain basic algorithms like search, sort etc. suited for various R data structures.
which_bisectSearch takes 2 main arguments: longDataVector, in which we will search for an index m, FOR WHICH 2nd argument, "sign_function" is equal to zero.
"sign_function" needs to be supplied by user, should return -1,0 or 1, and depends on longDataVector and index m and possible other values which are passed through  ... by the main function.
sign_function equals 0 if m is the index we are looking for. -1 if m is to small index, 1 if it's too big.

example sign_function is included